// class-components/src/app/interfaces/starships.interface.ts
export interface IStarships {
  MGLT: string;
  cargo_capacity: string;
  consumables: string;
  cost_in_credits: string;
  created: string;
  crew: string;
  edited: string;
  hyperdrive_rating: string;
  length: string;
  manufacturer: string;
  max_atmosphering_speed: string;
  model: string;
  name: string;
  passengers: string;
  films: string[];
  pilots: string[];
  starship_class: string;
  url: string;
}

export interface IStarshipsApi {
  count: number;
  next: string | null;
  previous: string | null;
  results: IStarships[];
}